dt-bindings: display: rockchip: dw-hdmi: use "ref" as clock name
"vpll" is a misnomer. A clock input to a device should be named after
the usage in the device, not after the clock that drives it. On the
rk3568 the same clock is driven by the HPLL.
This patch adds "ref" as a new alternative clock name for "vpll"

Changes since v3:
- Keep old clock name for compatibility reasons

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>