arm64: dts: rockchip: rk356x: Fix pcie2x1 address mapping
Arcane magic, but it fixes large storage corruption issues.

This mapping is designed to be more straightforward, inspired by
dt-bindings docs for sample pcie3x2 node:

  reg = <0x3 0xc0800000 0x0 0x390000>,
        <0x0 0xfe280000 0x0 0x10000>,
        <0x3 0x80000000 0x0 0x100000>;
  ranges = <0x81000000 0x0 0x80800000 0x3 0x80800000 0x0 0x100000>,
           <0x83000000 0x0 0x80900000 0x3 0x80900000 0x0 0x3f700000>;

I noticed that this is crafted so that there doesn't need to be
any translation other than dropping the high dword bits, and I
modified the ranges for pcie2x1 to follow the same principle.

Tested with > 16 GiB of TX and RX traffic using nvme.

I've also extended the config space size to 16 MiB, which is number
mentioned on wikipedia, which is my go to source for high quality
information.

Some of the previous ranges change attempts fixed the storage
corruption issues, but broke wifi pcie devices (which make use I/O
space, as opposed to my nvme).

With this change both nvme doesn't get corrupted, and pcie
wifi cards keep working... :)

Signed-off-by: Ondrej Jirman <megous@megous.com>